♻️ Streamline QDMI Qiskit plugin tests
