+++ test/python/plugins/qiskit/test_sampler.py
@@ -11,41 +11,30 @@
 from __future__ import annotations
 
 from collections import Counter
-from typing import TYPE_CHECKING
 
 import numpy as np
+import pytest
 from qiskit import QuantumCircuit
 from qiskit.circuit import ClassicalRegister, Parameter
-from qiskit.primitives import BaseSamplerV2
-from qiskit.primitives.containers import BitArray
 
-from mqt.core.plugins.qiskit import QDMISampler
-
-if TYPE_CHECKING:
-    from mqt.core.plugins.qiskit import QDMIBackend
+from mqt.core import fomac
+from mqt.core.plugins.qiskit import QDMIBackend, QDMISampler
 
 
-def _get_bit_array(data: object, register: str) -> BitArray:
-    """Extract a classical-register BitArray from sampler pub result data.
-
-    Returns:
-        The extracted bit array for the requested classical register.
-    """
-    bit_array = getattr(data, register, None)
-    assert isinstance(bit_array, BitArray)
-    return bit_array
-
-
-def test_sampler_instantiation(backend_with_mock_jobs: QDMIBackend) -> None:
-    """Sampler can be instantiated with a backend."""
-    sampler = QDMISampler(backend_with_mock_jobs)
-    assert isinstance(sampler, BaseSamplerV2)
-    assert sampler.backend == backend_with_mock_jobs
+@pytest.fixture
+def sampler() -> QDMISampler:
+    """Returns a QDMISampler based on the DDSIM backend."""
+    session = fomac.Session()
+    devices = session.get_devices()
+    for device in devices:
+        if "DDSIM" in device.name():
+            backend = QDMIBackend(device=device, provider=None)
+            return QDMISampler(backend)
+    pytest.skip("DDSIM device not available")
 
 
-def test_sampler_run_simple_circuit(backend_with_mock_jobs: QDMIBackend) -> None:
+def test_sampler_run_simple_circuit(sampler: QDMISampler) -> None:
     """Sampler runs a simple circuit."""
-    sampler = QDMISampler(backend_with_mock_jobs)
     qc = QuantumCircuit(2)
     qc.h(0)
     qc.cx(0, 1)
@@ -59,19 +48,18 @@ def test_sampler_run_simple_circuit(backend_with_mock_jobs: QDMIBackend) -> None
     assert pub_result.metadata["shots"] == 100
 
     # Check data bin structure
-    bit_array = _get_bit_array(pub_result.data, "meas")
+    bit_array = pub_result.data["meas"]
     assert bit_array.num_shots == 100
     assert bit_array.num_bits == 2
     assert bit_array.shape == ()
 
-    # Check that we got some counts (mock backend returns random distribution)
+    # Check that we got some counts
     counts = bit_array.get_counts()
     assert sum(counts.values()) == 100
 
 
-def test_sampler_run_parameterized_circuit(backend_with_mock_jobs: QDMIBackend) -> None:
+def test_sampler_run_parameterized_circuit(sampler: QDMISampler) -> None:
     """Sampler runs a parameterized circuit."""
-    sampler = QDMISampler(backend_with_mock_jobs)
     theta = Parameter("theta")
     qc = QuantumCircuit(1)
     qc.ry(theta, 0)
@@ -86,7 +74,7 @@ def test_sampler_run_parameterized_circuit(backend_with_mock_jobs: QDMIBackend) 
     assert pub_result.metadata["shots"] == 100
 
     # Shape should be (2,) because we provided 2 parameter sets
-    bit_array = _get_bit_array(pub_result.data, "meas")
+    bit_array = pub_result.data["meas"]
     assert bit_array.shape == (2,)
     assert bit_array.num_shots == 100
     assert bit_array.num_bits == 1
@@ -109,9 +97,8 @@ def test_sampler_run_parameterized_circuit(backend_with_mock_jobs: QDMIBackend) 
     assert sum(counts1.values()) == 100
 
 
-def test_sampler_run_multiple_cregs(backend_with_mock_jobs: QDMIBackend) -> None:
+def test_sampler_run_multiple_cregs(sampler: QDMISampler) -> None:
     """Sampler correctly handles multiple classical registers."""
-    sampler = QDMISampler(backend_with_mock_jobs)
     c0 = ClassicalRegister(1, "c0")
     c1 = ClassicalRegister(1, "c1")
     qc = QuantumCircuit(2)
@@ -125,41 +112,39 @@ def test_sampler_run_multiple_cregs(backend_with_mock_jobs: QDMIBackend) -> None
     result = job.result()
 
     pub_result = result[0]
-    c0_bits = _get_bit_array(pub_result.data, "c0")
-    c1_bits = _get_bit_array(pub_result.data, "c1")
+    c0_bits = pub_result.data["c0"]
+    c1_bits = pub_result.data["c1"]
 
     assert c0_bits.num_bits == 1
     assert c1_bits.num_bits == 1
 
 
-def test_sampler_options(backend_with_mock_jobs: QDMIBackend) -> None:
+def test_sampler_options(sampler: QDMISampler) -> None:
     """Test sampler options mechanism."""
     # 1. Use default shots from init
-    sampler = QDMISampler(backend_with_mock_jobs, default_shots=500)
+    sampler2 = QDMISampler(sampler.backend, default_shots=500)
     qc = QuantumCircuit(1)
     qc.measure_all()
 
-    job = sampler.run([(qc,)])
+    job = sampler2.run([(qc,)])
     result = job.result()
     assert result[0].metadata["shots"] == 500
 
     # 2. Override via run method
-    job = sampler.run([(qc,)], shots=200)
+    job = sampler2.run([(qc,)], shots=200)
     result = job.result()
     assert result[0].metadata["shots"] == 200
 
 
-def test_sampler_no_circuits(backend_with_mock_jobs: QDMIBackend) -> None:
+def test_sampler_no_circuits(sampler: QDMISampler) -> None:
     """Test run with empty pub list."""
-    sampler = QDMISampler(backend_with_mock_jobs)
     job = sampler.run([])
     result = job.result()
     assert len(result) == 0
 
 
-def test_sampler_broadcasting(backend_with_mock_jobs: QDMIBackend) -> None:
+def test_sampler_broadcasting(sampler: QDMISampler) -> None:
     """Test sampler with parameter broadcasting."""
-    sampler = QDMISampler(backend_with_mock_jobs)
     theta = Parameter("theta")
     qc = QuantumCircuit(1)
     qc.rx(theta, 0)
@@ -171,6 +156,6 @@ def test_sampler_broadcasting(backend_with_mock_jobs: QDMIBackend) -> None:
     result = job.result()
 
     pub_result = result[0]
-    bit_array = _get_bit_array(pub_result.data, "meas")
+    bit_array = pub_result.data["meas"]
     assert bit_array.shape == (2, 2)
     assert bit_array.num_shots == 100
